Attempt at trying to use header file for AppModLeds
