+++ src/SysModFiles.h
@@ -1,3 +1,4 @@
+#pragma once
 #include "Module.h"
 
 #include "LittleFS.h"
